Update the SPP to dual pathway.
